requirements.txt
certifi==2020.12.5
chardet==4.0.0
cryptocompare==0.7.5
# cycler==0.10.0
# idna==2.10
# kiwisolver==1.3.1
numpy==1.20.2
pandas==1.2.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
Flask==2.0.1
python-dotenv==0.17.1
ccxt==1.50.66
